requirements.txt
python-dotenv 
telethon
requests 
dvc
dbt-postgres
psycopg2-binary
ultralytics
fastapi
uvicorn
sqlalchemy
pydantic